+++ src/App.js
@@ -1,56 +1,57 @@
-import { useState } from "react";
-
-const messages = [
-  "Learn React ⚛️",
-  "Apply for jobs 💼",
-  "Invest your new income 🤑",
+const initialItems = [
+  { id: 1, description: " Passports", quantity: 2, packed: false },
+  { id: 2, description: " Socks", quantity: 12, packed: true },
+  { id: 3, description: " charger", quantity: 1, packed: false },
 ];
 
 export default function App() {
-  const [step, setStep] = useState(1);
-  const [isOpen, setIsOpen] = useState(true);
-
-  function handlePrevious() {
-    if (step > 1) setStep((s) => s - 1);
-  }
-  function handleNext() {
-    if (step < 3) {
-      setStep((s) => s + 1);
-      // setStep((s) => s + 1);
-    }
-  }
-
   return (
-    <>
-      <button className="close" onClick={() => setIsOpen((is) => !is)}>
-        &times;
-      </button>
-      {isOpen && (
-        <div className="steps">
-          <div className="numbers">
-            <div className={step >= 1 ? "active" : ""}>2</div>
-            <div className={step >= 2 ? "active" : ""}>1</div>
-            <div className={step >= 3 ? "active" : ""}>3</div>
-          </div>
-          <p className="message">
-            Step {step}: {messages[step - 1]}
-          </p>
-          <div className="buttons">
-            <button
-              style={{ backgroundColor: "#7950f2", color: "#fff" }}
-              onClick={handlePrevious}
-            >
-              Previous
-            </button>
-            <button
-              style={{ backgroundColor: "#7950f2", color: "#fff" }}
-              onClick={handleNext}
-            >
-              Next
-            </button>
-          </div>
-        </div>
-      )}
-    </>
+    <div className="app">
+      <Logo />
+      <Form />
+      <PackingList />
+      <Stats />
+    </div>
+  );
+}
+function Logo() {
+  return <h1>🌴 Far Away 💼</h1>;
+}
+function Form() {
+  return (
+    <div className="add-form">
+      <h3>What do you need for your 😍 trip?</h3>
+    </div>
+  );
+}
+function PackingList() {
+  return (
+    <div className="list">
+      <ul>
+        {initialItems.map((item, index) => (
+          <Item item={item} key={index} />
+        ))}
+      </ul>
+    </div>
+  );
+}
+
+function Item({ item }) {
+  return (
+    <li>
+      <span style={item.packed ? { textDecoration: "line-through" } : {}}>
+        {item.quantity}
+        {item.description}
+      </span>
+      <button>❌</button>
+    </li>
+  );
+}
+
+function Stats() {
+  return (
+    <footer className="stats">
+      <em>💼 You have X items on your list, and you already packed X (X%)</em>
+    </footer>
   );
 }
