+++ src/index.css
@@ -1,99 +1,131 @@
-/* 
-const messages = [
-  "Learn React ⚛️",
-  "Apply for jobs 💼",
-  "Invest your new income 🤑",
-];
- */
+@import url("https://fonts.googleapis.com/css2?family=Monoton&family=Quicksand:wght@500;700&display=swap");
 
 * {
+  margin: 0;
+  padding: 0;
   box-sizing: border-box;
 }
 
+html {
+  font-size: 62.5%;
+}
+
 body {
+  font-size: 2.4rem;
   font-family: sans-serif;
-  color: #333;
+  color: #5a3e2b;
+  font-family: "Quicksand";
+  font-weight: 500;
 }
 
-.steps {
-  width: 600px;
-  background-color: #f7f7f7;
-  border-radius: 7px;
-  padding: 25px 100px;
-  margin: 100px auto;
+.app {
+  width: 100%;
+  height: 100vh;
+  display: grid;
+  grid-template-rows: auto auto 1fr auto;
 }
 
-.numbers {
-  display: flex;
-  justify-content: space-between;
+h1 {
+  text-align: center;
+  background-color: #f4a226;
+  font-family: "Monoton";
+  font-size: 8rem;
+  text-transform: uppercase;
+  font-weight: 400;
+  word-spacing: 30px;
+  letter-spacing: -5px;
+  padding: 2.4rem 0;
 }
 
-.numbers > div {
-  height: 40px;
-  aspect-ratio: 1;
-  background-color: #e7e7e7;
-  border-radius: 50%;
+.add-form {
+  background-color: #e5771f;
+  padding: 2.8rem 0;
   display: flex;
   align-items: center;
   justify-content: center;
-  font-size: 18px;
-}
-
-.numbers .active {
-  background-color: #7950f2;
-  color: #fff;
-}
-
-.message {
-  text-align: center;
-  font-size: 20px;
-  margin: 40px 0;
-  font-weight: bold;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.buttons {
-  display: flex;
-  justify-content: space-between;
-}
-
-.buttons button {
-  border: none;
-  cursor: pointer;
-  padding: 10px 15px;
-  border-radius: 100px;
-  font-size: 14px;
-  font-weight: bold;
-
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.buttons button span {
-  font-size: 16px;
-  line-height: 1;
+  gap: 0.8rem;
 }
 
 h3 {
-  margin: 0;
-  text-transform: uppercase;
+  margin-right: 1.6rem;
+  font-size: 2.4rem;
 }
 
-.close {
-  position: absolute;
-  top: 16px;
-  right: 16px;
+button,
+select,
+input {
+  background-color: #ffebb3;
+  color: #5a3e2b;
+  font-family: inherit;
   border: none;
-  background: none;
+  border-radius: 10rem;
+  padding: 1.2rem 3.2rem;
+  font-weight: 700;
+  font-size: 1.8rem;
   cursor: pointer;
-  font-size: 40px;
-  color: inherit;
 }
 
-.close:hover {
-  color: #7950f2;
+.add-form button {
+  text-transform: uppercase;
+  background-color: #76c7ad;
+}
+
+.list {
+  background-color: #5a3e2b;
+  color: #ffebb3;
+  padding: 4rem 0;
+
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+  gap: 3.2rem;
+  align-items: center;
+}
+
+.actions button,
+.list select {
+  text-transform: uppercase;
+  padding: 0.8rem 2.4rem;
+  font-size: 1.4rem;
+  font-weight: 700;
+  margin: 0 0.8rem;
+}
+
+.list ul {
+  list-style: none;
+  width: 80%;
+  /* overflow: scroll; */
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+  gap: 1.2rem;
+  justify-content: center;
+  align-content: start;
+}
+
+.list li {
+  display: flex;
+  align-items: center;
+  gap: 1.2rem;
+}
+
+input[type="checkbox"] {
+  height: 2rem;
+  width: 2rem;
+  accent-color: #e5771f;
+}
+
+.list li button {
+  cursor: pointer;
+  background: none;
+  border: none;
+  font-size: 1.8rem;
+  padding: 0.8rem;
+  transform: translateY(2px);
+}
+
+.stats {
+  background-color: #76c7ad;
+  text-align: center;
+  font-weight: 700;
+  padding: 3.2rem 0;
 }
